+++ models/order/src/order-from-cart-draft/presets/sample-data-goodstore/index.ts
@@ -1,8 +1,8 @@
-import jenniferJonesCart01 from './jennifer-jones-cart';
-import sebastianSmithCart01 from './sebastian-smith-cart';
-import jenniferSchmidtCart01 from './jennifer-schmidt-cart';
-import sebastianMuellerCart01 from './sebastian-mueller-cart';
-import sebastianFranklinCart01 from './sebastian-franklin-cart';
+import jenniferJonesCart01 from './jennifer-jones-cart-01';
+import sebastianSmithCart01 from './sebastian-smith-cart-01';
+import jenniferSchmidtCart01 from './jennifer-schmidt-cart-01';
+import sebastianMuellerCart01 from './sebastian-mueller-cart-01';
+import sebastianFranklinCart01 from './sebastian-franklin-cart-01';
 
 const presets = {
   jenniferJonesCart01,
