+++ models/order/src/order-from-cart-draft/presets/sample-data-goodstore/jennifer-jones-cart-01.spec.ts
@@ -2,18 +2,18 @@ import {
   TOrderFromCartDraft,
   TOrderFromCartDraftGraphql,
 } from '../../../types';
-import sebastianFranklinCart01 from './sebastian-franklin-cart';
+import jenniferJonesCart01 from './jennifer-jones-cart-01';
 
-describe('with the preset order from cart draft `sebastianFranklinCart01`', () => {
+describe('with the preset order from cart draft `jenniferJonesCart01`', () => {
   it(`should create an order draft`, () => {
     const orderFromCartDraft =
-      sebastianFranklinCart01(1).build<TOrderFromCartDraft>();
+      jenniferJonesCart01(1).build<TOrderFromCartDraft>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 
   it(`should create an order draft when built for graphql`, () => {
     const orderFromCartDraft =
-      sebastianFranklinCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
+      jenniferJonesCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 });
