+++ models/order/src/order-from-cart-draft/presets/sample-data-goodstore/sebastian-franklin-cart-01.spec.ts
@@ -2,18 +2,18 @@ import {
   TOrderFromCartDraft,
   TOrderFromCartDraftGraphql,
 } from '../../../types';
-import jenniferSchmidtCart01 from './jennifer-schmidt-cart';
+import sebastianFranklinCart01 from './sebastian-franklin-cart-01';
 
-describe('with the preset order from cart draft `jenniferSchmidtCart01`', () => {
+describe('with the preset order from cart draft `sebastianFranklinCart01`', () => {
   it(`should create an order draft`, () => {
     const orderFromCartDraft =
-      jenniferSchmidtCart01(1).build<TOrderFromCartDraft>();
+      sebastianFranklinCart01(1).build<TOrderFromCartDraft>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 
   it(`should create an order draft when built for graphql`, () => {
     const orderFromCartDraft =
-      jenniferSchmidtCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
+      sebastianFranklinCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 });
