+++ models/order/src/order-from-cart-draft/presets/sample-data-goodstore/sebastian-franklin-cart-01.ts
@@ -6,18 +6,18 @@ import { TOrderFromCartDraftBuilder } from '../../../types';
 import * as OrderFromCartDraft from '../../index';
 
 const cart = CartDraft.presets.sampleDataGoodStore
-  .sebastianMuellerCart()
+  .sebastianFranklin01()
   .build<TCartDraft>();
 
-const sebastianMuellerCart01 = (
+const sebastianFranklinCart01 = (
   versionNumber: number
 ): TOrderFromCartDraftBuilder =>
   OrderFromCartDraft.presets
     .empty()
     .version(versionNumber)
     .cart(KeyReference.random().key(cart.key!).typeId('cart'))
-    .orderState(constants.orderState.Confirmed)
-    .paymentState(constants.paymentState.Paid)
-    .shipmentState(constants.shipmentState.Ready);
+    .orderState(constants.orderState.Open)
+    .paymentState(constants.paymentState.Pending)
+    .shipmentState(constants.shipmentState.Pending);
 
-export default sebastianMuellerCart01;
+export default sebastianFranklinCart01;
