+++ models/order/src/order-from-cart-draft/presets/sample-data-goodstore/sebastian-smith-cart-01.spec.ts
@@ -2,18 +2,18 @@ import {
   TOrderFromCartDraft,
   TOrderFromCartDraftGraphql,
 } from '../../../types';
-import sebastianMuellerCart01 from './sebastian-mueller-cart';
+import sebastianSmithCart01 from './sebastian-smith-cart-01';
 
-describe('with the preset order from cart draft `sebastianMuellerCart01`', () => {
+describe('with the preset order from cart draft `sebastianSmithCart01`', () => {
   it(`should create an order draft`, () => {
     const orderFromCartDraft =
-      sebastianMuellerCart01(1).build<TOrderFromCartDraft>();
+      sebastianSmithCart01(1).build<TOrderFromCartDraft>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 
   it(`should create an order draft when built for graphql`, () => {
     const orderFromCartDraft =
-      sebastianMuellerCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
+      sebastianSmithCart01(1).buildGraphql<TOrderFromCartDraftGraphql>();
     expect(orderFromCartDraft).toMatchInlineSnapshot();
   });
 });
